requirement.txt
numpy
pytorch==1.11.0
opencv-python
pyyaml
yacs
tqdm
colorama
matplotlib
cython
tensorboardX
